# modules/core-platform/api/v1/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    CreateInventoryAdjustmentRequest:
      type: object
      required:
      - variantId
      - locationId
      - quantityChange
      - reason
      - adjustedBy
      properties:
        variantId:
          $ref: "#/components/schemas/UUID"
        locationId:
          $ref: "#/components/schemas/UUID"
        quantityChange:
          type: integer
          format: int32
        reason:
          type: string
          pattern: \S
        adjustedBy:
          type: string
          pattern: \S
        notes:
          type: string
    CreateInventoryLocationRequest:
      type: object
      required:
      - code
      - name
      properties:
        code:
          type: string
          pattern: \S
          maxLength: 100
        name:
          type: string
          pattern: \S
          maxLength: 255
        type:
          type: string
          maxLength: 50
        address:
          type: string
        active:
          type: boolean
    CreateInventoryTransferRequest:
      type: object
      required:
      - sourceLocationId
      - destinationLocationId
      - lines
      properties:
        sourceLocationId:
          $ref: "#/components/schemas/UUID"
        destinationLocationId:
          $ref: "#/components/schemas/UUID"
        initiatedBy:
          type: string
        expectedArrivalDate:
          $ref: "#/components/schemas/OffsetDateTime"
        carrier:
          type: string
        trackingNumber:
          type: string
        shippingCost:
          type: integer
          format: int32
        notes:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/TransferLineRequest"
          minItems: 1
...
    OffsetDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50-04:00
...
    TransferLineRequest:
      type: object
      required:
      - variantId
      - quantity
      properties:
        variantId:
          $ref: "#/components/schemas/UUID"
        quantity:
          type: integer
          format: int32
        notes:
          type: string
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
paths:
...
  /api/v1/admin/inventory/adjustments:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryAdjustmentRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Create Adjustment
      tags:
      - Inventory Admin Resource
  /api/v1/admin/inventory/locations:
    get:
      responses:
        "200":
          description: OK
      summary: List Locations
      tags:
      - Inventory Admin Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryLocationRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Create Location
      tags:
      - Inventory Admin Resource
  /api/v1/admin/inventory/transfers:
    get:
      responses:
        "200":
          description: OK
      summary: List Transfers
      tags:
      - Inventory Admin Resource
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryTransferRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Create Transfer
      tags:
      - Inventory Admin Resource
  /api/v1/admin/inventory/transfers/{transferId}:
    get:
      parameters:
      - name: transferId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
      summary: Get Transfer
      tags:
      - Inventory Admin Resource
  /api/v1/admin/inventory/transfers/{transferId}/receive:
    post:
      parameters:
      - name: transferId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
      summary: Receive Transfer
      tags:
      - Inventory Admin Resource
...
info:
  title: Village Storefront API
  version: 1.0.0
  description: Multi-tenant SaaS ecommerce platform API
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value